requirements.txt

pandas
sqlalchemy
pymysql
sentence-transformers
faiss-cpu

